refactor: split DataService file into separate functions
