+++ src/interface/views/release/ConceptReport.vue
@@ -312,8 +312,8 @@ import error from "../../components/Error.vue";
 import * as d3 from "d3-time-format";
 import * as d3Format from "d3-format";
 import InfoPanel from "../../components/InfoPanel.vue";
-import dataService from "../../../services/DataService";
 import ReturnButton from "@/interface/components/ReturnButton";
+import sortByRange from "@/services/range-sort";
 
 export default {
   components: {
@@ -1248,7 +1248,7 @@ export default {
           }
 
           this.specRecordProportionByAgeSexYear.data = {
-            values: dataService.sortByRange(
+            values: sortByRange(
               this.conceptData.PREVALENCE_BY_GENDER_AGE_YEAR,
               "ascending",
               "TRELLIS_NAME",
