+++ src/services/get-percentage.js
@@ -1,0 +1,6 @@
+export default function getPercentage(value, array) {
+  const denominator = array.reduce((acc, current) => {
+    return acc + current.COUNT_VALUE;
+  }, 0);
+  return (value / denominator).toFixed(3) * 100 + " %";
+}
